Recover system cleanly on failed system start
When an error is thrown in a component's start phase, this can leave
some components in a started state without a system storing their
post-start state. This means that it can be impossible to stop these
components without restarting the JVM.

This commit solves this problem by catching exception on start and
recovering the partially started system by stopping components that have
started.